финальные мелкие изменения перед релизом

+++ SchedulePlanner/Controllers/SubjectController.cs
@@ -4,26 +4,40 @@ using SchedulePlanner.Db.Models;
 using SchedulePlanner.Db.Repositories;
 using System.Threading.Tasks;
 using System;
+using Microsoft.AspNetCore.Authorization;
+using System.Linq;
+using SchedulePlanner.ViewModels;
 
 namespace SchedulePlanner.Controllers
 {
+    [Authorize]
     public class SubjectController : Controller
     {
         private readonly ISubjectRepository _subjectRepository;
         private readonly UserManager<User> _userManager;
+        private readonly ILessonRepository _lessonRepository;
 
-        public SubjectController(ISubjectRepository subjectRepository, UserManager<User> userManager)
+        public SubjectController(ISubjectRepository subjectRepository, UserManager<User> userManager, ILessonRepository lessonRepository)
         {
             _subjectRepository = subjectRepository;
             _userManager = userManager;
+            _lessonRepository = lessonRepository;
         }
 
         public async Task<IActionResult> Index()
         {
             var user = await _userManager.GetUserAsync(User);
-            var periodId = user.SelectedPeriodId.GetValueOrDefault();
-            var subjects = _subjectRepository.GetByPeriodId(periodId);
-            return View(subjects);
+            if (user != null)
+            {
+                var periodId = user.SelectedPeriodId.GetValueOrDefault(); //Выбранный период
+                if (periodId == Guid.Empty)
+                {
+                    return RedirectToAction("Index", "Period");
+                }
+                var subjects = _subjectRepository.GetByPeriodId(periodId).Select(subject => SubjectViewModel.FromModel(subject));
+                return View(subjects);
+            }
+            return NotFound();
         }
 
         public async Task<IActionResult> Create()
@@ -36,20 +50,21 @@ namespace SchedulePlanner.Controllers
                 PeriodId = periodId
             };
             
-            return View(subject);
+            return View(SubjectViewModel.FromModel(subject));
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Subject subject)
+        public IActionResult Create(SubjectViewModel subjectViewModel)
         {
             if (ModelState.IsValid)
             {
+                var subject = subjectViewModel.ToModel();
                 subject.Id = Guid.NewGuid();
                 _subjectRepository.Add(subject);
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index");
             }
-            return View(subject);
+            return View(subjectViewModel);
         }
 
         public IActionResult Edit(Guid id)
@@ -59,24 +74,25 @@ namespace SchedulePlanner.Controllers
             {
                 return NotFound();
             }
-            return View(subject);
+            return View(SubjectViewModel.FromModel(subject));
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Edit(Guid id, Subject subject)
+        public IActionResult Edit(Guid id, SubjectViewModel subjectViewModel)
         {
-            if (id != subject.Id)
+            if (id != subjectViewModel.Id)
             {
                 return NotFound();
             }
 
             if (ModelState.IsValid)
             {
+                var subject = subjectViewModel.ToModel();
                 _subjectRepository.Update(subject);
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index");
             }
-            return View(subject);
+            return View(subjectViewModel);
         }
 
         public IActionResult Delete(Guid id)
@@ -86,15 +102,22 @@ namespace SchedulePlanner.Controllers
             {
                 return NotFound();
             }
-            return View(subject);
+            return View(SubjectViewModel.FromModel(subject));
         }
 
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public IActionResult DeleteConfirmed(Guid id)
         {
+            bool hasLesson = _lessonRepository.HasSubject(id);
+            if(hasLesson)
+            {
+                ModelState.AddModelError("", "Удаление невозможно, так как данный предмет используется в каком-то занятии");
+                var subject = _subjectRepository.GetById(id);
+                return View(SubjectViewModel.FromModel(subject));
+            }    
             _subjectRepository.Delete(id);
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Index");
         }
     }
 }
